refactor(groups): extract TaskForceEngine from GroupConversationService
R1 step 6 of planning/group-collaboration-improvements-plan.md -- the
largest and most concurrency-sensitive extraction in Wave R: the entire
TASK_FORCE PLAN/EXECUTE/VERIFY cluster (~840 lines) into a new
TaskForceEngine. This is the code GroupConversationServiceConcurrencyTest
exists to pin -- the taskList -> transcript lock order, the
recordTaskFailure/notifyTaskFailure split, and resetStrandedInProgressTasks's
compare-and-set-under-monitor sweep all moved verbatim.

Did the bare-token reflection sweep first this time (the step-5 lesson).
Found 12 of 16 methods reflected -- more than any prior step -- including
several reached via a direct multi-line getDeclaredMethod(...) call that a
single-line pattern grep can't see. All 12 kept as delegators; the other 4
(executeTaskPlanPhase, abortWave, stringOrNull, notifyTaskFailure) had zero
matches anywhere and were fully inlined.

A 13th reflected method, reserveTurn, was hiding just outside the cluster's
own banner (in the prior "cooperative cancellation" section) and is
reflected as a static invocation (invoke(null, ...)). Missed by the
line-range-scoped search, caught before compiling by checking every method
actually called from the code being moved rather than trusting banner
boundaries. Moved to TaskForceEngine with a static delegator left behind,
same pattern as propagateDynamicAgentTracking in step 4.

Two bugs caught before reaching CI: a transcription error where I invented
non-existent ...ForTest-suffixed delegator method names instead of matching
TaskForceEngine's real ones (caught re-reading the diff, not by the
compiler); and a stale-TaskItem bug in the new TaskForceEngineTest itself
(TaskItem is an immutable record -- completeTask() returns a new instance
rather than mutating the one already held locally, so a test was checking
status==COMPLETED against a stale PENDING snapshot).

Full 12-class group suite + DynamicAgentTrackingPropagationTest + 6 focused
collaborator test classes green (555 tests), including
GroupConversationServiceConcurrencyTest (8/8) and
GroupConversationServiceTaskForceTest (20/20).

GroupConversationService: 3,432 -> 2,594 lines. 6 of R1's 10 steps done;
GroupConversationService is now smaller than both AgentOrchestrator (2,725)
and ConversationService (2,698).
